requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dnspython==2.3.0
fonttools==4.39.3
frozenlist==1.3.3
idna==3.4
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
loguru==0.7.0
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.2
openai==0.27.4
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
pinecone-client==2.2.1
plotly==5.14.1
prompt-toolkit==1.0.14
pycodestyle==2.10.0
Pygments==2.15.1
PyInquirer==1.0.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
tenacity==8.2.2
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
yarl==1.8.2
zipp==3.15.0
